+++ ThePartingOfSarah/Background.h
@@ -8,7 +8,7 @@ public:
     Background(Game* game);
     Background(float x, float y, Game* game);
 
-    void draw(float scrollX = 0, float scrollY = 0) override;
+    void draw(int scrollX = 0, int scrollY = 0) override;
     void update();
 private:
     Background* bgAux = nullptr;
